# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Policy Orchestrator & Algorithm Execution API
  version: 1.0.0
  description: >-
    Enterprise-grade, vendor-agnostic REST API for repository invariant auditing,
    grounded hybrid GraphRAG, autonomous AI Agent orchestration, and direct execution
    of all 152 Layer 1 algorithms across Search, Observability, Update, Vector, Graph, Filter, and Transform categories.
    Strictly conforms to the RFC API Request & Response Structure Specification (v5.0).
servers:
  - url: http://localhost:8000
    description: Local development orchestrator
paths:
...
  /api/v1/algos/search/scan:
    post:
      summary: Multi-Pattern Fast Directory Scan (Search Algorithm Engine)
      operationId: scanDirectoryMultipattern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patterns]
              properties:
                root_dir:
                  type: string
                  default: "."
                patterns:
                  type: array
                  items:
                    type: string
                max_files:
                  type: integer
                  default: 1000
      responses:
        '200':
          description: Search algorithm match results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - requestId
        - correlationId
        - timestamp
        - apiVersion
        - status
        - trace_id
      properties:
        requestId:
          type: string
          example: "req-1700000000000-a1b2c3"
        correlationId:
          type: string
          example: "corr-1700000000000-x9y8z7"
        causationId:
          type: string
          nullable: true
        timestamp:
          type: string
          example: "2026-10-05T14:00:00.000Z"
        executionTimeMs:
          type: integer
          example: 12
        apiVersion:
          type: string
          example: "v1"
        status:
          type: string
          enum: [success, error]
        trace_id:
          type: string
          example: "0af7651916cd43dd8448eb211c80319c"
        auditLogId:
          type: string
          nullable: true
        operationId:
          type: string
          nullable: true
        deprecatedFields:
          type: array
          items: { type: string }
          nullable: true
        supportedVersions:
          type: array
          items: { type: string }
        staleness:
          type: object
          nullable: true
        pagination:
          type: object
          nullable: true
        links:
          type: object
          nullable: true
    ErrorContainer:
      type: object
      required:
        - code
        - message
        - retryable
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        retryable:
          type: boolean
          default: false
        details:
          type: array
          items:
            type: object
    ErrorItem:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        details:
          type: object
...
    Envelope_AlgoResponse:
      type: object
      required: [success, statusCode, meta, data]
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        data: { type: object }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { $ref: '#/components/schemas/ErrorContainer' }
        errors: { type: array, items: { $ref: '#/components/schemas/ErrorItem' }, nullable: true }
...